1.1 build 37: strict_types-Regressionen im ReceiveData-Pfad korrigiert
korrigiert:
- Segment: colorTempToRGB nimmt jetzt float entgegen — die Bisektion in RGBToColorTemp
  lieferte Bruchwerte und warf unter strict_types einen TypeError bei jedem Farbupdate
- Master: Nachtlicht-Restdauer (nl.rem) rechnet mit intdiv statt floor — DateTime::setTime
  verlangt unter strict_types int

Verhaltensgleichheit gegen build-35-Referenz verifiziert (rem identisch,
Farbtemperatur-Abweichung ≤ 0,7 K durch entfallene int-Quantisierung).

+++ SymconWLEDSegment/module.php
@@ -483,7 +483,7 @@ class WLEDSegment extends IPSModuleStrict
      *
      * @return array An array containing the red, green, and blue values as integers.
      */
-    private function colorTempToRGB(int $kelvin): array
+    private function colorTempToRGB(float $kelvin): array
     {
         $temp = $kelvin / 100;
